FAT MAJ : implémentation de load xml. Save a encore quelques problème, à ne pas utiliser pour le moment
la fonction load utilise la bibliothèque rapid xml, elle est pleinement fonctionnelle (enfin, selon les tests que je lui fais subir... je garantie rien xD par contre je fais le SAV en cas de bug) pour la partie projet, la partie Agenda est encore à implémenter.
Pour la faire fonctionner, il faut lui donner le chemin absolu (problème qui sera résolu plus tard)
Il y a l'exemple de fonctionnement dans le main !

Note pour le rapport : exportimport est une interface pour permettre l'implémentation d'autres méthodes d'import et d'export

Ajout d'une méthode virtuelle pure dans tacheUnitaire & Composite, on peut désormais se passer de la sublime méthode clone ^^
Modification de la classe Durée :
	ajout d'un constructeur avec pour paramètre un char *
		l'appel de Duree(0) n'est plus possible, il faut faire un static_cast<int> pour le résoudre.
	conversion d'une durée depuis un char *
Modification de la classe Evenement :
	ajout d'un "getID" nécessaire pour la sauvegarde de l'évent, la classe est maintenant abstraite
	modification de tache Unitaire en fonction

+++ Application/Classe/projetManager.h
@@ -3,6 +3,7 @@
 #include <vector>
 #include "projet.h"
 #include "tache.h"
+#include "duree.h"
 
 class ProjetManager
 {
@@ -42,7 +43,7 @@ public:
     TacheIterator tache_begin();
     TacheIterator tache_end();
     Projet* ajouterProjet(const QString &identificateur, const QString &ti, const QDate& dispo, const QDate& deadline);
-    Tache * ajouterTache(const QString & id_projet, const QString& id, const QString& titre, const QDate& dispo, const QDate& deadline, const Duree &dur = Duree(0), const bool &pre = false);
+    Tache * ajouterTache(const QString & id_projet, const QString& id, const QString& titre, const QDate& dispo, const QDate& deadline, const Duree &dur = Duree(static_cast<unsigned int>(0)), const bool &pre = false);
 
     std::vector<Projet *>& getProjets();
     const std::vector<Projet *>& getProjets() const; // Ne fonctionne pas. La surcharge n'est pas possible manifestement
